requirements.txt
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
click==8.0.4
decorator==5.1.1
executing==0.8.3
Flask==2.0.3
ipython==8.1.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.11.2
python-dotenv==0.20.0
six==1.16.0
stack-data==0.2.0
tomli==2.0.1
traitlets==5.1.1
typing_extensions==4.1.1
wcwidth==0.2.5
Werkzeug==2.0.3
